feat: add block three components related with testing
- Add testing components and suites
- Add mockery tooling in justfile
- Add environment files per use case
- Amend persistence layer in karma assignee module and remove in memory repo
- Add acceptance testing in the most basic use cases
- Add proper and identifiable name on each docker container declared on compose.yaml

+++ internal/karma-assignee/domain/karma_assignee.go
@@ -29,14 +29,14 @@ func NewKarmaAssigneeFromPrimitives(
 	id,
 	username,
 	assigner string,
-	count int64,
+	count uint64,
 	at time.Time,
 ) *KarmaAssignee {
 	return &KarmaAssignee{
 		id:        KarmaAssigneeId(id),
 		username:  Username(username),
 		assigner:  Username(assigner),
-		counter:   NewKarmaAssigneeCounterWithCount(uint64(count)),
+		counter:   NewKarmaAssigneeCounterWithCount(count),
 		createdAt: at,
 	}
 }
